fix: solve some changes on request in #665
