+++ src/components/UI/Header/index.tsx
@@ -60,19 +60,19 @@ export default function Header({
 
   if (renderBackButton) {
     return (
-      <div className={cx({ header: !title }, { headerSettings: title })}>
+      <header className={cx({ header: !title }, { headerSettings: title })}>
         <Link data-testid="returnLink" className="returnLink" to={link} onClick={handleClick}>
           <ArrowBack className="material-icons" />
           {t('Return')}
         </Link>
         {title && <div className="headerTitle" data-testid="headerTitle">{title}</div>}
-      </div>
+      </header>
     )
   }
 
   return (
     <>
-      <div className={cx({ header: !title }, { headerSettings: title })}>
+      <header className={cx({ header: !title }, { headerSettings: title })}>
         <span className="selectCategory">
           <span
             data-testid="gamesCategory"
@@ -209,7 +209,7 @@ export default function Header({
             onClick={() => handleLayout('list')}
           ></List>
         </div>
-      </div>
+      </header>
     </>
   )
 }
