+++ src/screens/FreeProducts/utils/testUtils.tsx
@@ -1,4 +1,4 @@
-import { Element } from './apiResponseType';
+import { FreeGameElement } from 'src/types'
 
 /**
  * Reset the promotion dates for 2 preselected products
@@ -6,7 +6,7 @@ import { Element } from './apiResponseType';
  * @param elements    List of products with promotions from the Mock API response
  * @returns           No return value as object is altered inline
  */
-export const resetDates = (elements: Element[]) => {
+export const resetDates = (elements: FreeGameElement[]) => {
   const validProducts = ['d89d1ecf209d42688d82909e522f2ec1', '0c7924b21589422581c7cbe8d5a336c5']
   elements.forEach(element => {
     if (element.promotions && validProducts.includes(element.id) ){
